requirements.txt
streamlit
pydub
moviepy
google-cloud-speech
google-cloud-texttospeech
openai
requests
tempfile